[fx] Replace literals with placeholder helper
Helper function to replace literals that show up in call_function nodes
in the graph to become placeholders so that they can be represented as
wildcards when matching with the SubgraphMatcher. This pass causes the
resulting graph to not be runnable with the original inputs since adding
placeholders to the graph will change the number of inputs needed for
the graph.

Test: `python test/test_fx.py TestMatcher`