[WIP][profiler] initialize kineto before doing any profiling.
As suggested in robieta in #75504, this workaround seems to fix the cudagraph profiling issues. It moves the kineto initialization step to torch initialization, so that it gets called before any cuda graphs are created.

Tested locally and verified that the issue goes away (although - locally, the symptom is a hanging process, not an illegal memory access).

ghstack-source-id: 8c61769f69bc4ea9e43300f07ffce2a73be95c43
Pull Request resolved: https://github.com/pytorch/pytorch/pull/97199